extending estimation for nested calls WIP
