+++ src/logic/safe/transactions/gas.js
@@ -1,59 +1,5 @@
 // @flow
 import { getGnosisSafeInstanceAt } from '~/logic/contracts/safeContracts'
-import { EMPTY_DATA } from '~/logic/wallets/ethTransactions'
-
-const estimateDataGasCosts = (data) => {
-  const reducer = (accumulator, currentValue) => {
-    if (currentValue === EMPTY_DATA) {
-      return accumulator + 0
-    }
-
-    if (currentValue === '00') {
-      return accumulator + 4
-    }
-
-    return accumulator + 68
-  }
-
-  return data.match(/.{2}/g).reduce(reducer, 0)
-}
-
-// https://docs.gnosis.io/safe/docs/docs4/#safe-transaction-data-gas-estimation
-// https://github.com/gnosis/safe-contracts/blob/a97c6fd24f79c0b159ddd25a10a2ebd3ea2ef926/test/utils/execution.js
-export const estimateDataGas = (
-  safe: any,
-  to: string,
-  valueInWei: number,
-  from: string,
-  data: string,
-  operation: number,
-  txGasEstimate: number,
-  gasToken: number,
-  nonce: number,
-  signatureCount: number,
-  refundReceiver: number,
-) => {
-  // numbers < 256 are 192 -> 31 * 4 + 68
-  // numbers < 65k are 256 -> 30 * 4 + 2 * 68
-  // For signature array length and dataGasEstimate we already calculated
-  // the 0 bytes so we just add 64 for each non-zero byte
-  const gasPrice = 0 // no need to get refund when we submit txs to metamask
-  const signatureCost = signatureCount * (68 + 2176 + 2176 + 6000) // array count (3 -> r, s, v) * signature count
-
-  // https://docs.gnosis.io/safe/docs/docs5/#pre-validated-signatures
-  const sigs = `0x000000000000000000000000${from.replace(
-    '0x',
-    '',
-  )}000000000000000000000000000000000000000000000000000000000000000001`
-  const payload = safe.contract.methods
-    .execTransaction(to, valueInWei, data, operation, txGasEstimate, 0, gasPrice, gasToken, refundReceiver, sigs)
-    .encodeABI()
-
-  // eslint-disable-next-line
-  const dataGasEstimate = estimateDataGasCosts(payload) + signatureCost + (nonce > 0 ? 5000 : 20000) + 1500 // 1500 -> hash generation costs
-
-  return dataGasEstimate + 32000 // Add aditional gas costs (e.g. base tx costs, transfer costs)
-}
 
 export const calculateTxFee = async (
   safe: any,
